products/serializer/products.py
from rest_framework import serializers
from products.models.products import Product

class ProductSerializer(serializers.ModelSerializer):
  class Meta:
    model = Product
    fields=  '__all__'
    